add geo class
