+++ config.py
@@ -35,3 +35,19 @@ def remove_column(column_name):
     global META_DATA
     if column_name in META_DATA:
         META_DATA.remove(column_name)
+
+month_abbr_hydr = {
+    'sep': 1,
+    'oct': 2,
+    'nov': 3,
+    'dec': 4,
+    'jan': 5,
+    'feb': 6,
+    'mar': 7,
+    'apr': 8,
+    'may': 9,
+    'jun': 10,
+    'jul': 11,
+    'aug': 12,
+    #'sep_': 13,
+}
