+++ massbalancemachine/__init__.py
@@ -1,8 +1,8 @@
 import os, sys; sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 
-__all__ = ['dataloader', 'models', 'data_processing', 'plot']
+__all__ = ['dataloader', 'models', 'data_processing', 'geodata']
 
 from .dataloader import *
 from .models import CustomXGBoostRegressor, utils
 from .data_processing import Dataset, utils
-from .plot import *
+from .geodata import *
