+++ massbalancemachine/geodata/__init__.py
@@ -1,0 +1,3 @@
+import os, sys; sys.path.append(os.path.dirname(os.path.realpath(__file__)))
+
+from massbalancemachine.geodata.GeoData import GeoData
